+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project (attempts to) adhere to [Semantic Versioning](http://semver.org/).
 
+## [1.2.4] - 2020-02-27
+- No longer crash when "N" is in the reference sequence
+- Add the ability to --use_full_fasta_header
+
 ## [1.2.3] - 2020-02-21
 - Messed with the internals of filter_reads. Moved old methods into deprecated_filter_reads
 - Changed the options of filter_reads to just do things that it can actually do at the moment